dashboard/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
Flask-Session==0.6.0
Werkzeug==3.0.1
gunicorn==21.2.0
whitenoise==6.6.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
discord.py==2.3.2
aiohttp==3.9.3 